Adrig Hackathon/streamlitapp/requirements.txt
streamlit
pandas
matplotlib
seaborn
scikit-learn
plotly-express
PIL